@@ -1,105 +1,79 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * This file is part of the SimplexCalculator library
  *
  * Copyright (c) 2014 Petr Kessler (http://kesspess.1991.cz)
  *
  * @license  MIT
+ *
  * @link     https://github.com/uestla/Simplex-Calculator
  */
 
 namespace Simplex;
 
-
 class Restriction extends VariableSet
 {
 
-	/** @var int */
-	private $type;
+    public const TYPE_EQ = 1;
+    public const TYPE_LOE = 2;
+    public const TYPE_GOE = 4;
+    private int $type;
 
-	/** @var Fraction */
-	private $limit;
+    private Fraction $limit;
 
+    public function __construct(array $set, int $type, $limit)
+    {
+        parent::__construct($set);
 
+        $this->type = (int)$type;
+        $this->limit = Fraction::create($limit);
+    }
 
-	const TYPE_EQ = 1;
-	const TYPE_LOE = 2;
-	const TYPE_GOE = 4;
+    /** Deep copy */
+    public function __clone()
+    {
+        $this->limit = clone $this->limit;
+    }
 
+    public function getType(): int
+    {
+        return $this->type;
+    }
 
+    public function getLimit(): Fraction
+    {
+        return $this->limit;
+    }
 
-	/**
-	 * @param  array $set
-	 * @param  int $type
-	 * @param  Fraction|numeric $limit
-	 */
-	function __construct(array $set, $type, $limit)
-	{
-		parent::__construct($set);
+    public function fixRightSide(): Restriction
+    {
+        if ($this->limit->isLowerThan(0)) {
+            $set = [];
+            foreach ($this->set as $var => $coeff) {
+                $set[$var] = $coeff->multiply(-1);
+            }
 
-		$this->type = (int) $type;
-		$this->limit = Fraction::create($limit);
-	}
+            $type = $this->type === self::TYPE_EQ ? $this->type
+                : ($this->type === self::TYPE_GOE ? self::TYPE_LOE : self::TYPE_GOE);
 
+            $this->limit = $this->limit->multiply(-1);
+        } else {
+            $set = $this->set;
+            $type = $this->type;
+        }
 
+        $this->set = $set;
+        $this->type = $type;
 
-	/** @return int */
-	function getType()
-	{
-		return $this->type;
-	}
+        return $this;
+    }
 
-
-
-	/** @return Fraction */
-	function getLimit()
-	{
-		return $this->limit;
-	}
-
-
-
-	/** @return Restriction */
-	function fixRightSide()
-	{
-		if ($this->limit->isLowerThan(0)) {
-			$set = array();
-			foreach ($this->set as $var => $coeff) {
-				$set[$var] = $coeff->multiply(-1);
-			}
-
-			$type = $this->type === self::TYPE_EQ ? $this->type
-					: ($this->type === self::TYPE_GOE ? self::TYPE_LOE : self::TYPE_GOE);
-
-			$this->limit = $this->limit->multiply(-1);
-
-		} else {
-			$set = $this->set;
-			$type = $this->type;
-		}
-
-		$this->set = $set;
-		$this->type = $type;
-
-		return $this;
-	}
-
-
-
-	/** @return string */
-	function getTypeSign()
-	{
-		return $this->type === self::TYPE_EQ ? '='
-				: ($this->type === self::TYPE_LOE ? "\xe2\x89\xa4" : "\xe2\x89\xa5");
-	}
-
-
-
-	/** Deep copy */
-	function __clone()
-	{
-		$this->limit = clone $this->limit;
-	}
-
+    public function getTypeSign(): string
+    {
+        return $this->type === self::TYPE_EQ ? '='
+            : ($this->type === self::TYPE_LOE ? "\xe2\x89\xa4" : "\xe2\x89\xa5");
+    }
 }
